Placed TLS Configuration builders into core/config/config.go, which makes more sense than having them in an own file. Thus removed unnecessary files.
