+++ Gopkg.lock
@@ -114,6 +114,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "79eb333130a261ad8daca59ba54fbd92bc6206a0d7520723f899bd7a1fc1fd06"
+  inputs-digest = "b5834a148c9fd7a62319ea9a01fb4c6b593da4d27daf0c86067aa110297d71e2"
   solver-name = "gps-cdcl"
   solver-version = 1
